+++ partiql-lang/src/main/kotlin/org/partiql/lang/eval/ExprFunction.kt
@@ -17,6 +17,9 @@ package org.partiql.lang.eval
 import org.partiql.lang.errors.ErrorCode
 import org.partiql.lang.types.FunctionSignature
 
+@Deprecated("As function overloading is now supported, this class is deprecated.", level = DeprecationLevel.ERROR)
+sealed class Arguments
+
 /**
  * Represents a function that can be invoked from within an [EvaluatingCompiler]
  * compiled [Expression].
@@ -43,6 +46,18 @@ interface ExprFunction {
      * @param required The required arguments.
      */
     fun callWithRequired(session: EvaluationSession, required: List<ExprValue>): ExprValue {
+        errNoContext("Invalid implementation for ${signature.name}#call", ErrorCode.INTERNAL_ERROR, true)
+    }
+
+    @Deprecated("As function overloading is now supported, please provide multiple implementations of ExprFunction and invoke callWithRequired() instead.", level = DeprecationLevel.ERROR)
+    fun callWithOptional(session: EvaluationSession, required: List<ExprValue>, opt: ExprValue): ExprValue {
+        // Deriving ExprFunctions must implement this if they have a valid call form including required parameters and optional
+        errNoContext("Invalid implementation for ${signature.name}#call", ErrorCode.INTERNAL_ERROR, true)
+    }
+
+    @Deprecated("As function overloading is now supported, please use the LIST ExprValue to represent variadic parameters and invoke callWithRequired() instead.", level = DeprecationLevel.ERROR)
+    fun callWithVariadic(session: EvaluationSession, required: List<ExprValue>, variadic: List<ExprValue>): ExprValue {
+        // Deriving ExprFunctions must implement this if they have a valid call form including required parameters and variadic
         errNoContext("Invalid implementation for ${signature.name}#call", ErrorCode.INTERNAL_ERROR, true)
     }
 }