+++ CHANGELOG.md
@@ -43,15 +43,22 @@ Thank you to all who have contributed!
 ### Changed
 - Standardizes `org/partiql/cli/functions/QueryDDB` and other built-in functions by the new `ExprFunction` format.
 - **Breaking**: Redefines `org/partiql/lang/eval/ExprFunctionkt.call()` method by only invoking `callWithRequired` function.
-- **Breaking**: Redefines `org/partiql/lang/eval/builtins/DynamicLookupExprFunction` by merging `variadicParameter` into `requiredParameters` as a `StaticType.LIST`. `callWithVariadic` is now replaced by`callWithRequired`.
+- **Breaking**: Redefines `org/partiql/lang/eval/builtins/DynamicLookupExprFunction` by merging `variadicParameter` into `requiredParameters` as a `StaticType.LIST`. `callWithVariadic` is now replaced by `callWithRequired`.
 - **Breaking**: Modifies `functions` property of `CompilerPipeline`, `StaticTypeInferencer` and  `StepContext` to be a `List` and not a `Map`.
 
 ### Deprecated
+- Deprecates Map<String, ExprFunction> representation of functions in the `CompilerPipeline`
+  and experimental `PartiQLCompilerPipeline`. Please use List<ExprFunction> to represent functions instead.
+- **Breaking**: Deprecates `Arguments` class, `callWithOptional()` and `callWithVariadic()` methods in the `ExprFunction` 
+  with a Deprecation Level of ERROR. Please invoke `callWithRequired()` instead.
+- **Breaking**: Deprecates `optionalParameter` and `variadicParameter` in the `FunctionSignature` with a Deprecation 
+  Level of ERROR. Please use multiple implementations of ExprFunction and use the LIST ExprValue to
+  represent variadic parameters instead.
 
 ### Fixed
 
 ### Removed
-- **Breaking**: Removes `optionalParameter` and `variadicParameter` from`org.partiql.lang.types.FunctionSignature`. To continue support for evaluation of `optionalParameters`, please create another same-named function. To continue support for evaluation of `variadicParameter`, please use a `StaticType.LIST` to hold all previously variadic parameters.
+- **Breaking**: Removes `optionalParameter` and `variadicParameter` from `org.partiql.lang.types.FunctionSignature`. To continue support for evaluation of `optionalParameters`, please create another same-named function. To continue support for evaluation of `variadicParameter`, please use a `StaticType.LIST` to hold all previously variadic parameters.
   As this changes coincides with the addition of function overloading, only `callWithRequired` will be invoked upon execution of an `ExprFunction`.
 - **Breaking**: Removes unused class `Arguments` from `org.partiql.lang.eval`.
 - **Breaking**: Removes unused parameter `args: Arguments` from `org.partiql.lang.eval.ExprFunctionkt.call()` method.